Trigger actions on push as well (#4)
This is necessary so the action is run on merges into the main branch
too. That wouldn't necessarily be required since we enforce the checks
to succeed before merging anyway, but we want to run our actions on
main too in order to populate the caches that can then be reused on PR
branches. Otherwise, each PR branch would have a first run that doesn't
have any caches available.

https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#restrictions-for-accessing-a-cache